requirements.txt
autopep8==1.2.1
numpy==1.10.4
Pillow==3.1.0
pydicom==0.9.9
scipy==0.17.0
